requirements.txt
pyTelegramBotAPI
gTTS
python-dotenv